[extension] trigger script injection & data reconnects from background
This is necessary for single-page apps like slack where switching
between channels will change the URL but not reload the page. In those
cases we need to be notified of location changes (via
`chrome.tabs.onUpdated`) and possibly inform the toolbar that it
needs to reconnect to subscribe to a different puzzle's data.