[extension] Move data fetching from popup to background/index.js
This makes opening the popup much quicker, as the background is
already maintaining a firebase connection. The initial auth flow
also takes place in the background script now, which means that
the success callback will run to completion. Previously, the popup
would close as soon as the oauth window opens, which means that
nothing happens after authentication is finished.